Former-commit-id: 7afe360fff4b1cbf57782aa4b84e36cf892077c7

+++ src/data/MAIN_Process_10_31.java
@@ -3,9 +3,11 @@ package data;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Map.Entry;
+
+import evaluators.Evaluator_HandTunedOnState;
+import game.State;
+
 import java.util.TreeMap;
-import main.Evaluator_HandTunedOnState;
-import main.State;
 
 /**
  * Processes the data from 10/31/17
@@ -136,7 +138,7 @@ public class MAIN_Process_10_31 {
 					System.out.println("Warning: loaded a null game");
 				}
 				allLoadedRuns.add(game);
-				System.out.println((int)((j+1.)/(float)inFiles.size() * 100) + "% total completion. " + (int)((j - k)/(float)fileBatchSize * 100) + "% of file batch imported.");
+				System.out.println((int)((j+1.)/inFiles.size() * 100) + "% total completion. " + (int)((j - k)/(float)fileBatchSize * 100) + "% of file batch imported.");
 			}
 			System.out.printf("Done\n");
 
@@ -187,7 +189,7 @@ public class MAIN_Process_10_31 {
 
 			ProcessDenseData reSim = new ProcessDenseData();
 			reSim.process(allSortedGames, gameBatchSize);
-			System.out.println("Completed file batch " + (int)(k/fileBatchSize + 1) + "/" + (int)Math.ceil(inFiles.size()/(double)fileBatchSize));
+			System.out.println("Completed file batch " + (k/fileBatchSize + 1) + "/" + (int)Math.ceil(inFiles.size()/(double)fileBatchSize));
 
 		}
 	}
